fix permission error when adding backports
I think those instructions are too complicated - `sh -c` will be difficult to parse for new (or even experienced) users and is error prone. Nevertheless, the instructions as is will not work because the redirection does not run as root, as is the case for the other instructions. This was probably forgotten when adding the backports entry (which goes to show this is error-prone). (I would instead recommend "piping to tee", like `echo ... | sudo tee -a sources.list`, for example.)